+++ src/RightTurn.Extensions.Configuration/TurnConfigurationExtensions.cs
@@ -1,16 +1,20 @@
 ﻿using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using System;
-using System.IO;
 
 namespace RightTurn.Extensions.Configuration
 {
     public static class TurnConfigurationExtensions
     {
+        public static ITurn WithServices(this ITurn turn, Action<IConfiguration, IServiceCollection> services)
+        {
+            turn.Directions.Add<ITurnServices>(new TurnServices(services));
+            return turn;
+        }
+
         public static ITurn WithConfiguration(this ITurn turn, Func<IConfiguration> builder)
         {
-            turn.Directions.Add(builder);
-            turn.Directions.Add<ITurnConfiguration>(new TurnConfiguration());
+            turn.Directions.Add<ITurnConfiguration>(new TurnConfiguration(builder));
             return turn;
         }
 
@@ -21,8 +25,7 @@ namespace RightTurn.Extensions.Configuration
                 .Build();
         });
 
-        public static ITurn WithConfigurationFile(this ITurn turn, string name = "appsettings.json", bool optional = true) => WithConfiguration(turn, (builder) => builder
-            .SetBasePath(Directory.GetCurrentDirectory())
+        public static ITurn WithConfigurationFile(this ITurn turn, string name = "appsettings.json", bool optional = true) => WithConfiguration(turn, (builder) => builder            
             .AddJsonFile(name, optional));
 
 
@@ -36,7 +39,7 @@ namespace RightTurn.Extensions.Configuration
         /// <returns></returns>
         public static ITurn WithConfigurationSettings<T>(this ITurn turn, string key) where T : class, new()
         {
-            TurnConfiguration.Bindings.Add(key, (services) =>
+            turn.Directions.Add<TurnConfigurationBindings>().Add(key, (services) =>
             {
                 var settings = new T();
                 services.AddSingleton(settings);
@@ -56,7 +59,7 @@ namespace RightTurn.Extensions.Configuration
         /// <returns></returns>
         public static ITurn WithConfigurationSettings<TService, TImplementation>(this ITurn turn, string key) where TService : class where TImplementation : class, TService, new()
         {
-            TurnConfiguration.Bindings.Add(key, (services) =>
+            turn.Directions.Add<TurnConfigurationBindings>().Add(key, (services) =>
             {
                 var service = new TImplementation();
                 services.AddSingleton<TService>(service);