+++ CHANGELOG.md
@@ -2,6 +2,20 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## [0.2.0](https://github.com/Jandini/RightTurn.Extensions.Configuration/compare/0.1.0...0.2.0) (2021-03-27)
+
+
+### Features
+
+* Add access to configuration when adding services ([49276ce](https://github.com/Jandini/RightTurn.Extensions.Configuration/commit/49276cedebc8e481eb66c892f6530cdb4aa5c4a0))
+* Add configuration required exception ([2b7beb8](https://github.com/Jandini/RightTurn.Extensions.Configuration/commit/2b7beb8aa39a098ea8ee5a45341f82f4ba34955e))
+* Improve configuration extensions and binding ([94c21bd](https://github.com/Jandini/RightTurn.Extensions.Configuration/commit/94c21bd79755d77f99fc3d1b1cd691a4b1136736))
+
+
+### Bug Fixes
+
+* Application published as single file executable can access appsettings.json file ([1c03cad](https://github.com/Jandini/RightTurn.Extensions.Configuration/commit/1c03cad34b96dfa089055300edf70d846ed1650b))
+
 ## 0.1.0 (2021-03-21)